+++ packages/ethereum-contracts/contracts/agreements/ConstantFlowAgreementV1.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: AGPLv3
 pragma solidity ^0.8.23;
 
-import { CallbackUtils } from "../libs/CallbackUtils.sol";
 import {
     ISuperfluid,
     ISuperfluidGovernance,
@@ -1049,7 +1048,6 @@ contract ConstantFlowAgreementV1 is
     // solhint-disable-next-line contract-name-camelcase
     struct _StackVars_changeFlowToApp {
         bytes cbdata;
-        uint256 remainingCallbackGas;
         FlowData newFlowData;
         ISuperfluid.Context appContext;
     }
@@ -1086,8 +1084,7 @@ contract ConstantFlowAgreementV1 is
             } else /* if (optype == FlowChangeType.DELETE_FLOW) */ {
                 cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_TERMINATED_NOOP;
             }
-            (vars.cbdata, vars.remainingCallbackGas) = AgreementLibrary.callAppBeforeCallback(
-                cbStates, CallbackUtils.HOST_CALLBACK_GAS_LIMIT, ctx);
+            (vars.cbdata, newCtx) = AgreementLibrary.callAppBeforeCallback(cbStates, ctx);
 
             ISuperfluidGovernance gov = ISuperfluidGovernance(ISuperfluid(msg.sender).getGovernance());
 
@@ -1125,8 +1122,7 @@ contract ConstantFlowAgreementV1 is
             } else /* if (optype == FlowChangeType.DELETE_FLOW) */ {
                 cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_TERMINATED_NOOP;
             }
-            (vars.appContext, newCtx) = AgreementLibrary.callAppAfterCallback(
-                cbStates, vars.cbdata, vars.remainingCallbackGas, newCtx);
+            (vars.appContext, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
 
             // NB: the callback might update the same flow!!
             // reload the flow data